requirements.txt
PyMySQL==1.1.0
ColorInfo==2.2.0
pycryptodome==3.19.0
pandas==2.1.4